requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
pytest==7.4.3
pytest-django==4.7.0
setuptools==80.9.0
dj-database-url==2.1.0
